deepseek_finetune/requirements.txt
transformers>=4.36.0
datasets>=2.14.0
torch>=2.1.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.4
wandb>=0.16.0
sentencepiece>=0.1.99
protobuf>=4.25.1
gradio>=4.13.0
huggingface_hub>=0.19.0
PyYAML>=6.0
requests>=2.26.0 